Feature #19119 - Allow OpenAPI 3 assertion - Added helper class to load resources - Moved test resources to test folder
