+++ dina-test-support/src/test/resources/ma-response-example.txt
