+++ dina-test-support/src/test/resources/managedAttribute.yaml
